feat(Typescript factory functions): Only first required prop is unnamed
This change attempts to balance the convienience of unnamed arguments, with the robustness of named arguments (aka keyword arguments in Python and other langs), particularly to changes in the number and order of required properties. It should mean that consumers of these factory functions do not need to change their code as often, when/if there are changes to type schemas.